requirements.txt
faiss-cpu
sentence-transformers
numpy
fastapi
uvicorn
python-telegram-bot==13.15
openai
urllib3
google-generativeai
google